Creating Tutorial
End goal is to build multiple pop ups that are linked together to walk
the user through the app.

Found this example,
http://stackoverflow.com/questions/30967057/different-content-in-multipl
e-css-popups
not sure it will work though.  Sort of tried to use it before I left,
but it just made it so you only could see the first comment and
couldn’t close it anymore.